Refactor socket into modules (similar to file), for -layer.
