+++ sample/rust/Cargo.toml
@@ -7,3 +7,4 @@ edition = "2021"
 
 [dependencies]
 libc = "0.2"
+reqwest = { version = "0.11", features = ["blocking"] }
